WebServer: Handle keep-alive connections
Browsers use "Connection: keep-alive".

The original definition in RFC 2608 is "Connection: Keep-Alive" for compatibility with HTTP/1.0 specified in RFC 1945 which has no mention of "Connection" or "Keep-Alive".

So, we have to handle both of these possibilities.

And I thought the rot of the web was limited to whatever garbage spewed out of HTML. Why does HTTP also have to suck?!