Allocate HTTP clients via a new StaticPool facility
A long-running HTTP client does not need to hold onto memory pertaining to processing a request when it is done.
Also, an HTTP client running as a WebSocket session does not need to be holding onto HTTP response state when the connection is established.
Having pools for these and potentially holding these back until they're available will allow better reuse of RAM.