JETTISON LIBEPOXY AND USE OPENGL DIRECTLY
One non-system dependency down, one to go.

I had to write an XML parser for this. Wasn't actually as bad as I thought it would be.
Unfortunately, for the schema I do need to improve a lot of stuff, so currently I've got some manual massaging in there.
And it's incomplete, so I need to massage it further.

Also, no C# was harmed in the making of this wrapper.

FUTURE PROPOSAL: Actually generate this at compile time without massaging and with minimal special-casing in the API gen.
FUTURE PROPOSAL: Walk the XML structure without allocating it into RAM because that gets really slow.
FUTURE PROPOSAL: A native type-safe Zig wrapper.