win32: Get keyboard input working
The Windows build is now at feature-parity with the Linux SDL2 build, and also notably ahead of the Windows SDL2 build as that wouldn't even link lol

Next step is to work out why the web server crashes - I think I'm getting some garbage in the high 32 bits of a 64-bit value which should probably be a 32-bit value. Go figure.