win32: This might actually work on Windows Vista now
I have a Windows 2000 JP VM. So I thought I'd try to get it working on that. Well, I failed in that regard, but, y'know, someone might still want to use Windows 7...

I rolled back to a bunch of older APIs and the like. This does include going to RegisterClassW which brings that part from 2000 to I think 95 and NT 4. However, the brick wall I hit was AcquireSRWLockExclusive, added in Windows Vista and used in a lot of std.Thread things.

Working around this would most likely require modifying the Zig standard library, which is not something I'm willing to do in the name of retrocomputing. Sorry, this is as far as I'm willing to push it back.

I may reinstate WSAPoll as an option if it's available in ws2_32.dll. Or alternatively I could use newer-style APIs for better performance... and Linux could do with epoll and/or io_uring, and if someone steps up and does a FreeBSD port and/or a macOS port then we could use kqueue for that.