Start sorting out the licensing mess
After actually reading the licence, we're going with AGPLv3 for the engine and also the overall distribution, and any assets and code we make which run inside the engine will be CC0. Talk about extremes lol

There are a few third-party things with their own licences, so I have to sift through all that stuff... and I also need to regenerate the final composed font for GNU Unifont and, due to it most likely counting as a modified font, slap a name on it (which will most likely be "sekaigu yunifon", and quite frankly we're probably going to add something to the PUA space for a munted う glyph - it has too much cultural significance, but I'm not mangling the glyph in-place as I would like to enjoy having, well, a non-munted う).