Finally get all the licensing stuff sorted! (well, as good as we can get it anyway)
Closes GH-1

This involved making a tool to generate our modified Unifont, which turns out it spat out exactly what I think I did using VIM, sort and uniq or whatever it was I did. Anyway, we now have a tool which generates it at build time.

I'll split off the Unifont attribution problem into a separate issue. There's a change we'll end up replacing it with public domain fonts if this turns out to be a pain.